[IMP] l10n_es_reports: Implement LIBRO DIARIO XLSX report export
Objective
---------
Libro Diario is just a big name for "Journal" in Spanish, but it's also
a precise report that is expected to reflect a chronological report of
all Accounting Entries in a Database. It's expected to be published and
available to show all the transactions recorded by a Company.

It's basically an export of the General Ledger with all the accounting
entries, not grouped by Account but by Accounting Entries, sorted by
date, and to show a full sequence.

The Libro Diario should posess the following columns and each row should
contain an account move line:
- "Entry": ascending numbering for each move, starting at 1.
- "Line": numbering of the move line inside the move, starting at 1. 
- "Date": accounting date of the move; as mentionned, moves should be
sorted by dates
- "Description": label of the move line
- "Document": account move name
- "Account code"
- "Account name"
- "Debit"
- "Credit"

Solution
---------
1. Add a custom General Ledge Handler that adds the option to export the
Libro Diarios
2. Since `_get_line` returns the AMLs grouped by accounts, it is easier
to fetch the query that gets all the AMLs data (`_get_query_amls`) and
then postprocess the amls data to generate the matrix of data passed to
the XLSX builder.

task-4294391

closes odoo/enterprise#75418

X-original-commit: 77ca24776a47aaf1fd22593b9616bdcd5d128005
Signed-off-by: Ruben Gomes (rugo) <rugo@odoo.com>