[FIX] account_followup: added a 2nd default level in data
The second level already configured in demo has been moved in data because having only one follow-up level in a new database is a bit weird. Not only does it look weird, but in practice it also leads to a weird situation that by default the invoices will only be reminded once and never after. + we're failing to convey the usefuleness of the feature of building gradually harded reminders for the partner.

Also added forcecreate="False" in snailmail demo data, because we don't want to recreate the missing xml ids (if people deleted it) in every module's update.

task-4334758

closes odoo/enterprise#75026

X-original-commit: 84aa5e2fb064bc0366e952cbd4eaf3628ea24b73
Signed-off-by: William André (wan) <wan@odoo.com>