[IMP] l10n_uk_{hmrc,reports_cis}: Secure way of rendering xml for hmrc_transaction_request
We now render templates from file instead of the database.
This way, the user cannot change the template to be rendered to HMRC and send malicious data or anything else.

This also adds more flexibility with the stable policy as we don't need to pay attention when changing file in stable.

closes odoo/enterprise#74193

Related: odoo/odoo#187933
Related: odoo/upgrade#6956
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>