[FIX] documents_spreadsheet: ensure consistency of folder's company
```
Disclaimer: The below description has been written for Odoo 17. It
may not be that simple to set a company on a folder with the classic
front-end. Still, the use case can be reproduced with custo/RPC.
```

There are several issues with the company of the folders and the
field `documents_spreadsheet_folder_id` of a company.

To reproduce the issue 01:
1. Create a second company
2. Documents > Configuration > Workspaces, edit "Spreadsheet":
   - Set a company

Step 01, `documents_spreadsheet_folder_id` of the second company is
set to "Spreadsheet" folder thanks to a default value on the field.
But, looking at the DB after the second step, there is an error: both
companies still have the folder "Spreadsheet" defined on
`documents_spreadsheet_folder_id`, but this folder is now specific to
one company. Step 2 should not be possible, a constraint is missing.

To reproduce the issue 02:
1. Repeat above steps 1-2
2. Create a third company

Looking at the DB, there is also an error with the third company: the
default value has been used for it, even though the folder is now
company-specific and can't be used anymore. The default value should
be improved to handle this situation.

Also, it should not be possible to use such record on the field.
Since there is already the attribute `check_company=True` on
`documents_spreadsheet_folder_id`, we could therefore add the attribute
`_check_company_auto` on the model and take advantage of the ORM:
https://github.com/odoo/odoo/blob/f323896028114acb36848b605732797b67fd7da4/odoo/models.py#L614-L618
However, this change would be too invasive on stable, hence the ORM
constraint as alternative.

OPW-4281530

closes odoo/enterprise#75284

X-original-commit: 8512948bf46d9d69be8360f2056e143c401b158f
Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>
Signed-off-by: Adrien Widart (awt) <awt@odoo.com>