[IMP] l10n_{be,sa}_hr_payroll: move reason code to locas
Remove the `reason_code` from `hr.hr_departure_reason` because
it is only relevant in certain localizations.

Replace with one field per localization where it is relevant, this is
needed because the code is not necessarily the same in every
localization that uses it.

task-4320913

closes odoo/enterprise#73729

Related: odoo/odoo#187046
Related: odoo/upgrade#6801
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>