[IMP] knowledge: attachments on the first comment
This commit enables the user to create a comment thread by sending a message
containing an attachment. Previously it wasn't possible to upload an attachment
to the 1st comment of a thread as we didn't know what the thread's id
would've been.
Now, when uploading an attachment on a new thread, the thread record is
effectively created. If no message is sent in the end, the thread will be
cleaned the next time the article is opened by the author.

Task-4331677

closes odoo/enterprise#74871

Signed-off-by: Damien Abeloos (abd) <abd@odoo.com>
Co-authored-by: Damien Abeloos <abd@odoo.com>
Co-authored-by: Kadam Darji <kmdi@odoo.com>
Co-authored-by: Thomas Josse <thjo@odoo.com>