[FIX] timesheet_grid: fix the timer header in list view
Currently, the timesheet header is not working correctly in the list
view. There are a bunch of issue that this commit aims to fix:

An empty header is displayed when the start button is pressed, and
trying to either stop it or discard it will triggers an error

The grouped by function triggers a similar issue as the one above when
trying to start & stop the timer

When there are no timesheetable project inside the database, an error is
triggered too when starting the timer.

Source of the issues :
- when trying to start a timer in the list view, the function
'cleartimesheetname' is triggered. The problem is that function triggers
an update on the listview. This update then triggers a second execution
of the 'onWillUpdateProps' of the timesheet_timer_hook class. This
second execution is not expected in the flow and set the value of the
timesheet of the header to 'false'.

- when the view is grouped, the record are not loaded inside the the
folded section. This means that when we search for the record inside the
'popRecord' function of the timesheet_timer_hook class, we find nothing.
In this use case though, the record is supposed to be found and set.

- when there are no timesheetable project, the method 'popRecord' will
set the timesheet of the header back to 'undefined'.

Solution :

- Removed the 'cleartimesheetname' method, as it is no longer usefull. A
check up in the write/create of the model will prevent the required
field 'description' from being null. The 'enterEditMode' has also been
restricted to the kanban view, since doing this in the list view was
also triggering an extra call to 'onWillupdateprops'

- Checked the view to get a dynamicRecordList we can use to correctly
set the data. More information inside the comment in the code.

- added a condition inside the 'popRecord' method so that in no longer
set the timesheet to undefined when we dont need it to

task - 4268687
affected version 18.0-master

closes odoo/enterprise#75460

X-original-commit: ece12f502b1cace94f351afddeb14a7a3d17a4cd
Related: odoo/odoo#190292
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>