[IMP] planning: review avatar card layout
This commit reviews the layout of the avatar card in order to match all
the use cases.

The employee avatar card recently received an update aiming to display
skills and roles within its content. While this added extra information
and value to the component, its layout wasn't really meant to contain
that much information, which led to a crowded visual result.

With this commit, we review the hierarchy of the card by adding clear
separations to its content and reviewing the use of white space.

task-3810438

closes odoo/enterprise#74279

Related: odoo/odoo#188068
Signed-off-by: Chrysanthe Gomrée (chgo) <chgo@odoo.com>