[IMP] knowledge: file box style
This commit, alongside changes in html_editor, changes the style of the
file box (result of the /file command).

Formerly a banner, the new file card has a more compact style and no
longer has a download button. The new style allows for grid display of
multiple files in the same line.

task-3522395

X-original-commit: 1e09a3c507e4a06356ff70ce737917d9b9b5fd5f
Part-of: odoo/enterprise#77530
Related: odoo/odoo#194609
Signed-off-by: David Monjoie (dmo) <dmo@odoo.com>