[FIX] web_studio: update menu editor dialog when modifying a menu
In studio, click on Edit Menu in the main navbar.
Add a menu or edit the name of a menu.
Click Confirm

Before this commit, the hierarchical list of the menu editor was not updated.
This was due to the fact that studio's dialog service is altered to allow it to have the specific studio sub environment
It is that way since 2ae63d9a43ba973d905232105201d88e423a8245
As a consequence, the "menu changed" event triggerd by the menu service on the main environment's bus was not forwarded to to studio's
env's bus.

After this commit, the hierarchical list is updated correctly

task-4001101

closes odoo/enterprise#74516

X-original-commit: bbd1be5653899c20685f9f722cdb27f2fbde82de
Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>