[IMP] l10n_in_reports_gstr: match pos in gstr2b
In GSTR-2B matching, we only match the total amount,
not the IGST, CGST, and SGST which could lead to inaccuracy.
These taxes depend upon the Place of supply, so this commit adds
place of supply in matching, and if the place of supply
does not match then an exception message is shown on the bill.

task-4343860

Part-of: odoo/enterprise#74472
Signed-off-by: Josse Colpaert (jco) <jco@odoo.com>