[FIX] project_mrp_workorder_account: add profitability test
This commit's purpose is to add a test for the full flow of the creation
of a BOM/MO when they are linked to a project. This is to ensure that
the aal generated are generated correctly and that the project
profitability is then computing them into the correct section.

task - 4184226
affetced version 18.0 - master

closes odoo/enterprise#76381

X-original-commit: 6687517a89af014ebf64b4c37271c893d3c8bb2c
Related: odoo/odoo#192182
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>