[FIX] quality_mrp_workorder: workorder list view fixes
This is only a cosmetic change from a text button to an icon button.

The main changes of the PR are done in https://github.com/odoo/odoo/pull/184308/.

-----------------------
Community pr description:

- changed the "start", "pause" and "done" button text to icons to be able to
set the button column size at a fixed value
- moved the unblock button in the dropdown
- made the previous block button toggle between block and unblock
- change the dropdown bullet to red when blocked

This commit also modifies the width of the json_popover as it was using
the base column width which was too big for the icon.

It also modifies the width of the mo_list_dropdown column to limit the empty space.
This implies changes in the web module:
in `column_width_hook.js`, the width of a widget is equal to it's `listViewWidth`,
like the fields. However, unlike the fields, the widget class is missing a validation
for that props.

closes odoo/enterprise#74828

Task-id: 4279789
X-original-commit: 8451187613ebe981f7bf87a8d62247a3a020ea2f
Related: odoo/odoo#189061
Signed-off-by: Steve Van Essche <svs@odoo.com>