[IMP] iot: add checkbox to skip printer selection
Currently, you only get to select a printer the
first time you print a report (that has IoT printers
linked). The only way to change the printer is to
use the 'Reset Linked Printers' button in the IoT
app.

This PR changes the behaviour to always show the
selection popup, unless the user ticks a checkbox
to not show it again. The selection auto-fills to
whatever was selected last time for the report
(saved in local storage).

task-4129000

closes odoo/enterprise#74558

Signed-off-by: Yaroslav Soroko (yaso) <yaso@odoo.com>