[FIX] mrp_workorder: simpler relevant records
The idea is:

1. Stop to do a lot of thing in the getter `relevantRecords` because
   all those operations are done every time this getter is called.
   To achieve that, relevant records are computed each time before the
   component should be re-rendered and the getter only returns the
   computed result.
   It was making the code very hard to work with and very unstable.

2. Move `MrpDisplaySearchModel` `recordCache.ids` into `MrpDisplay`
   `recordCacheIds`. Why ? Because this array is always used in
   `MrpDisplay` so it doesn't do a lot of sense to define it in
   `MrpDisplaySearchModel`. The only exception is in the `_notify` where
   the cache is invalidated but it does not seem needed anymore ?
   Also, instead of having an object with one uniq key who refers to an
   array (`recordCache.ids`), it will now be an array directly.

Also, this commit fixes an issue where the record cache wasn't correctly
updated when the selected employee requieres a PIN code.
The issue was the cache was invalided before the PIN code was validated,
a lot of async methods weren't awaited and the dialog itself wasn't
bound to a callback in a way to be able to invalidate cache at the right
moment. A test is added to ensure this flow.

OPW-4256568

closes odoo/enterprise#79281

X-original-commit: 399598aca6b79b6c0f54fe5473135ff1e151ce90
Signed-off-by: Arnold Moyaux (arm) <arm@odoo.com>
Signed-off-by: Steve Van Essche <svs@odoo.com>