[IMP] hr_work_entry_holidays: prevent conflicting work entries.
Previously, if you created a leave request without approving
it or if the leave request required 2 approvers and it was
only approved the first time, then there would be a conflicting
work entry.However, this was changed by the community PR #188172
and there are no conflicting work entries now in that case.

This commit updates a testcase to match the new behavior.

task-4342340

closes odoo/enterprise#74426

Related: odoo/odoo#188172
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>