[FIX] quality_control: check correct model
The test tests that some controller returns the right data, however the
route was using the wrong model (and consequently id).
The test was working properly by chance with demo data (ids seemed to be
the same for both template and quality check spreadsheet).

Runbot Error 104754

Part-of: odoo/enterprise#74427
Related: odoo/odoo#188323
Signed-off-by: William Braeckman (wbr) <wbr@odoo.com>