[IMP] documents_sign: remove sign base folder
We remove the base folder configuration for sign. That folder was used as base
folder when creating a folder on the fly in the sign app. It's still possible
to create a folder on the fly in the sign app but the parent folder is not
configurable anymore. The parent folder will always be the "Sign" folder
created when installing documents_sign. In order that this folder remains
visible for all company, we ensure that its field company_id remains False. We
also ensure that this base folder cannot be deleted.

Note: as sign_template has no company_id, we don't create company specific
folder when creating a folder from a sign template.

We also change the wording of the link to the sign document templates in the
settings.

Task-4294237

Part-of: odoo/enterprise#74571
Related: odoo/odoo#191459
Related: odoo/upgrade#6820
Signed-off-by: Stéphane Debauche (std) <std@odoo.com>