[FIX] l10n_mx_edi_landing: add customs emission date to invoice lines
To comply with Mexican tax regulations (Article 29-A, section VII of the Fiscal Code and the CFDI guide, pg. 22), both the customs number and its emission date must be included in product descriptions for national sales.

This commit adds the date(s) of the emission of the customs number(s) to every customer invoice line label

task-4132754

closes odoo/enterprise#75503

X-original-commit: 21b89b0f8bf4bf565b635d7a17ea4b461a517a23
Signed-off-by: Laurent Smet (las) <las@odoo.com>
Signed-off-by: Anton Romanova (roan) <roan@odoo.com>