[IMP] whatsapp: auto-open chat window on new message
Currently, when receiving a new message, it opens a new chat bubble,
i.e. a chat window that is folded on the right.

This is the preferred behaviour for most users. For whatsapp
operators, however, they want to reply to visitors questions as
quickly as possible. Therefore having the chat window open
automatically on receiving new messages is quite useful for them.

This PR adds automatically opening of chat window on receiving new
messages for whatsapp conversations. All other conversations are
still open as chat bubble, as this is the preferred showing for all
other users that want to know there are new messages without being
disturbed while using Odoo web client.

Task-4354021

closes odoo/enterprise#74547

X-original-commit: 54e1ae09077b2c43ee611b84f9aa500a5562d667
Related: odoo/odoo#188536
Signed-off-by: Alexandre Kühn (aku) <aku@odoo.com>