[IMP] documents_project: add Project Folder to company settings
We add a settings allowing the user to choose the folder where new project
folders will appear for a company.

Notes:
- res_company.documents_project_folder_id is not made required as the default
folder doesn't exist when the field is added, but we set it as required in the
settings form and ensure that it is set on every company in the post_init.

- a constraint ensures that the base project folder remains open to all company
(company_id = False) as it is the default value for new companies.

Task-4294237

Part-of: odoo/enterprise#74571
Related: odoo/odoo#191459
Related: odoo/upgrade#6820
Signed-off-by: Stéphane Debauche (std) <std@odoo.com>
Co-authored-by: Florian Charlier <flch@odoo.com>