[IMP] website_appointment: allow to customize avatar number
Prior to this commit, the maximum number of avatars displayed in the
`appointment_booking_resource_description` template was set to 6.
After this commit, we can now customize the amount of avatars displayed,
simply by defining the `avatars_max_count` variable, to match the layout
where this template is called.

task-4317830

Part-of: odoo/enterprise#74494
Related: odoo/odoo#190946
Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>