[FIX] documents: expose groupless embedded actions
**Current behavior:**
Having some embedded action for `documents.document` which has
no `groups_ids` set will result in that action not ever
appearing in the actions bar/menu in the documents app.

**Expected behavior:**
In the absence of any records in the `groups_ids` relation, the
actions should be exposed globally.

**Steps to reproduce:**
*`-i documents_account` for example*
1. In technical settings, open the embedded actions list view

2. Edit the `Create Vendor Bill` action so that it has an empty
`groups_ids` field

3. Open the documents app, click on the finance tab on the left
sidebar

4. Click on one of the demo documents, observe in the actions
menu for the record that the create vendor bill action is no
longer present

**Cause of the issue:**
The domain was constructed incorrectly, as the
`('field', 'in', [False] + X.ids)`
pattern does not work when `'field'` is itself a m2m recordset.

**Fix:**
Split this domain leaf into 2 so we can check for membership as
well as falsy-ness.

opw-4342841

closes odoo/enterprise#75387

X-original-commit: f99692d0bc9cf19791eadf04cdc3b13aed0ff728
Signed-off-by: Vincent Ethan <etvi@odoo.com>