[IMP] room: Autocomplete scss rules cleanup
There are some rules for the ui-autocomplete elements define on some
scss files inside the /web/static/src/legacy folder linked to the
jQuery-UI autocomplete dropdown.

This commit cleans up these applied rules and remove it if they are not
currently needed. We also plan to remove these classes in the future

task-4330695

closes odoo/enterprise#74697

Related: odoo/odoo#188812
Signed-off-by: Adrien Dieudonné (adr) <adr@odoo.com>