[FIX] account_loans: allow zero interest loans
Problem:
Users are not allowed to set a zero interest rate on a loan.

Steps to reproduce:
1. Go to accounting accounting > loans > new
2. Set any name and any amount in the borrowed field
3. Click on Calculate
4. Set interest rate to 0% and click Apply

Explication:
The library we were using "loanpy" does not handle the case where the interest rate is zero if the loan_type is `annuity`, and we ended up dividing zero by zero, throwing an error.

The fix:
If the interest is zero, we override the loan's type from `annuity` (the default) to `linear`, as the linear formula is better suited for the zero interest case.

opw-4298272

closes odoo/enterprise#74668

X-original-commit: 42b7b322d6c720de45661e3370397adaf74fe1f4
Signed-off-by: Quentin De Paoli (qdp) <qdp@odoo.com>
Signed-off-by: Hadi El Yakhni (hael) <hael@odoo.com>