[IMP] web_studio: use new nice urls in studio
Before this commit, studio had its own navigation url. In particular, it meant that studio had
no trace whatsoever of the "breadcrumbs stack" represented in the url.

After this commit, studio adds itself at the end of that stack, keeping the full path all the way.
The edited action is always the one just preceding studio in the stack.

part of task-4356704

Part-of: odoo/enterprise#74800
Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>