[FIX] l10n_mx_reports: display partners infos on all DIOT lines
Steps to reproduce:
- Install "l10n_mx" and switch to a Mexican company
- Create a vendor bill with this partner using the tax "Exento"
- Go to Reports -> Tax report -> DIOT.txt

Cause:
For partners that only have lines with the tax "Exento", no information are retrieved for the report. The columns "Type of third", "type of operation", "VAT", etc. are empty in the Tax report before trying to download the DIOT report.
When trying to generate the txt file, the code tries to access the value "third_party_code" in a dictionnary but the key is not there so an error is raised.

This occurs because the tax "Exento" is "on_invoice" so there are no cash basis entry created. But in the custom query to retrieve the information of the partners it is specified the journal_id must be part of the cash basis journals. So no lines with "on_invoice" taxes were selected.

Solution:
Remove the journal restriction.

opw-4312155

closes odoo/enterprise#74740

X-original-commit: d78d00eb24f07ac39610141880e23753bcca0113
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>
Signed-off-by: Mathieu Coutant (mcou) <mcou@odoo.com>