[FIX] account_base_import: Don't skip tests if openpyxl installed
Since `xlrd >= 2.0` dropped XLSX support, we use `openpyxl` to open
XLSX files if the `xlrd >= 2.0` is installed.[^1]

However, the tests in `account_base_import` are skipped unless
`xlrd.xlsx` can be imported. As a result, they are not run on runbot,
where `xlrd >= 2.0` is installed.

We therefore need to avoid skipping them if openpyxl is installed.

runbot-108001

[^1]: https://github.com/odoo/odoo/pull/169245

closes odoo/enterprise#75290

X-original-commit: 0eb887668345e90688ca5405aabd2a6fb2eb6954
Signed-off-by: Antoine Dupuis (andu) <andu@odoo.com>
Signed-off-by: Xavier Morel (xmo) <xmo@odoo.com>