[IMP] sale_subscription: copy followers on renew or upsell order
Before this commit, when a renewal or upsell order was created, the followers
from the parent subscription were not carried over. Users had to manually add
these followers to the new order.

In this commit, followers from the parent subscription are automatically copied
to the renewal or upsell orders.

task-4355082

closes odoo/enterprise#74754

Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>